requirements.txt
django ~= 2.2
django-modeltranslation ~= 0.13
pillow ~= 6.0.0
